Plan 9 from Bell Labs 2005-09-01

+++ sys/src/cmd/spin/main.c
@@ -1,14 +1,13 @@
 /***** spin: main.c *****/
 
-/* Copyright (c) 1991-2000 by Lucent Technologies - Bell Laboratories     */
+/* Copyright (c) 1989-2003 by Lucent Technologies, Bell Laboratories.     */
 /* All Rights Reserved.  This software is for educational purposes only.  */
-/* Permission is given to distribute this code provided that this intro-  */
-/* ductory message is not removed and no monies are exchanged.            */
-/* No guarantee is expressed or implied by the distribution of this code. */
-/* Software written by Gerard J. Holzmann as part of the book:            */
-/* `Design and Validation of Computer Protocols,' ISBN 0-13-539925-4,     */
-/* Prentice Hall, Englewood Cliffs, NJ, 07632.                            */
-/* Send bug-reports and/or questions to: gerard@research.bell-labs.com    */
+/* No guarantee whatsoever is expressed or implied by the distribution of */
+/* this code.  Permission is given to distribute this code provided that  */
+/* this introductory message is not removed and no monies are exchanged.  */
+/* Software written by Gerard J. Holzmann.  For tool documentation see:   */
+/*             http://spinroot.com/                                       */
+/* Send all bug-reports and/or questions to: bugs@spinroot.com            */
 
 #include <stdlib.h>
 #include "spin.h"
@@ -19,6 +18,9 @@
 #ifdef PC
 #include <io.h>
 #include "y_tab.h"
+
+extern int unlink(const char *);
+
 #else
 #include <unistd.h>
 #include "y.tab.h"
@@ -28,18 +30,26 @@ extern int	DstepStart, lineno, tl_terse;
 extern FILE	*yyin, *yyout, *tl_out;
 extern Symbol	*context;
 extern char	*claimproc;
+extern void	repro_src(void);
 extern void	qhide(int);
 
 Symbol	*Fname, *oFname;
 
-int	Etimeouts=0;	/* nr timeouts in program */
-int	Ntimeouts=0;	/* nr timeouts in never claim */
-int	analyze=0, columns=0, dumptab=0, has_remote=0;
-int	interactive=0, jumpsteps=0, m_loss=0, nr_errs=0;
-int	s_trail=0, ntrail=0, verbose=0, xspin=0, no_print=0, no_wrapup = 0, Caccess=0;
-int	limited_vis=0, like_java=0;
-int	dataflow = 1, merger = 1, deadvar = 1, rvopt = 0, ccache = 1;
-int	separate = 0;	/* separate compilation option */
+int	Etimeouts;	/* nr timeouts in program */
+int	Ntimeouts;	/* nr timeouts in never claim */
+int	analyze, columns, dumptab, has_remote, has_remvar;
+int	interactive, jumpsteps, m_loss, nr_errs, cutoff;
+int	s_trail, ntrail, verbose, xspin, notabs, rvopt;
+int	no_print, no_wrapup, Caccess, limited_vis, like_java;
+int	separate;	/* separate compilation */
+int	export_ast;	/* pangen5.c */
+int	inlineonly;	/* show inlined code */
+int	seedy;		/* be verbose about chosen seed */
+
+int	dataflow = 1, merger = 1, deadvar = 1, ccache = 1;
+
+static int preprocessonly, SeedUsed;
+
 #if 0
 meaning of flags on verbose:
 	1	-g global variable values
@@ -51,8 +61,6 @@ meaning of flags on verbose:
 	64	-w very verbose
 #endif
 
-static Element	*Same;
-static int	IsAsgn = 0, OrIsAsgn = 0;
 static char	Operator[] = "operator: ";
 static char	Keyword[]  = "keyword: ";
 static char	Function[] = "function-name: ";
@@ -68,27 +76,35 @@ static void	explain(int);
 		/* OS2:		"spin -Picc -E/Pd+ -E/Q+"    */
 		/* Visual C++:	"spin -PCL  -E/E             */
 #ifdef PC
-#define CPP	"cpp"		/*  or specify a full path    */
+#define CPP	"gcc -E -x c"	/* most systems have gcc anyway */
+				/* else use "cpp" */
 #else
 #ifdef SOLARIS
 #define CPP	"/usr/ccs/lib/cpp"
 #else
-#ifdef __FreeBSD__
+#if defined(__FreeBSD__) || defined(__OpenBSD__)
 #define CPP	"cpp"
 #else
-#define CPP	"/bin/cpp"
+#define CPP	"/bin/cpp"	/* classic Unix systems */
 #endif
 #endif
 #endif
 
 #endif
 static char	*PreProc = CPP;
+extern int	depth; /* at least some steps were made */
 
 void
 alldone(int estatus)
 {
-	if (strlen(out1) > 0)
+	if (preprocessonly == 0
+	&&  strlen(out1) > 0)
 		(void) unlink((const char *)out1);
+
+	if (seedy && !analyze && !export_ast
+	&& !s_trail && !preprocessonly && depth > 0)
+		printf("seed used: %d\n", SeedUsed);
+
 	if (xspin && (analyze || s_trail))
 	{	if (estatus)
 			printf("spin: %d error(s) - aborting\n",
@@ -101,10 +117,10 @@ alldone(int estatus)
 
 void
 preprocess(char *a, char *b, int a_tmp)
-{	char precmd[128], cmd[256]; int i;
+{	char precmd[512], cmd[1024]; int i;
 #ifdef PC
 	extern int try_zpp(char *, char *);
-	if (try_zpp(a, b)) goto out;
+	if (PreCnt == 0 && try_zpp(a, b)) goto out;
 #endif
 	strcpy(precmd, PreProc);
 	for (i = 1; i <= PreCnt; i++)
@@ -115,9 +131,13 @@ preprocess(char *a, char *b, int a_tmp)
 	if (system((const char *)cmd))
 	{	(void) unlink((const char *) b);
 		if (a_tmp) (void) unlink((const char *) a);
-		alldone(1); /* failed */
+		fprintf(stdout, "spin: preprocessing failed\n");	/* 4.1.2 was stderr */
+		alldone(1); /* no return, error exit */
 	}
-out:	if (a_tmp) (void) unlink((const char *) a);
+#ifdef PC
+out:
+#endif
+	if (a_tmp) (void) unlink((const char *) a);
 }
 
 FILE *
@@ -142,10 +162,11 @@ usage(void)
 {
 	printf("use: spin [-option] ... [-option] file\n");
 	printf("\tNote: file must always be the last argument\n");
+	printf("\t-A apply slicing algorithm\n");
 	printf("\t-a generate a verifier in pan.c\n");
 	printf("\t-B no final state details in simulations\n");
 	printf("\t-b don't execute printfs in simulation\n");
-	printf("\t-C print channel access info (structview)\n");
+	printf("\t-C print channel access info (combine with -g etc.)\n");
 	printf("\t-c columnated -s -r simulation output\n");
 	printf("\t-d produce symbol-table information\n");
 	printf("\t-Dyyy pass -Dyyy to the preprocessor\n");
@@ -155,7 +176,9 @@ usage(void)
 	printf("\t-F file  like -f, but with the LTL ");
 	printf("formula stored in a 1-line file\n");
 	printf("\t-g print all global variables\n");
+	printf("\t-h  at end of run, print value of seed for random nr generator used\n");
 	printf("\t-i interactive (random simulation)\n");
+	printf("\t-I show result of inlining and preprocessing\n");
 	printf("\t-J reverse eval order of nested unlesses\n");
 	printf("\t-jN skip the first N steps ");
 	printf("in simulation trail\n");
@@ -165,16 +188,20 @@ usage(void)
 	printf("\t-N file use never claim stored in file\n");
 	printf("\t-nN seed for random nr generator\n");
 	printf("\t-o1 turn off dataflow-optimizations in verifier\n");
-	printf("\t-o2 turn off dead variables elimination in verifier\n");
+	printf("\t-o2 don't hide write-only variables in verifier\n");
 	printf("\t-o3 turn off statement merging in verifier\n");
 	printf("\t-Pxxx use xxx for preprocessing\n");
 	printf("\t-p print all statements\n");
 	printf("\t-qN suppress io for queue N in printouts\n");
 	printf("\t-r print receive events\n");
+	printf("\t-S1 and -S2 separate pan source for claim and model\n");
 	printf("\t-s print send events\n");
+	printf("\t-T do not indent printf output\n");
+	printf("\t-t[N] follow [Nth] simulation trail\n");
+	printf("\t-Uyyy pass -Uyyy to the preprocessor\n");
+	printf("\t-uN stop a simulation run after N steps\n");
 	printf("\t-v verbose, more warnings\n");
 	printf("\t-w very verbose (when combined with -l or -g)\n");
-	printf("\t-t[N] follow [Nth] simulation trail\n");
 	printf("\t-[XYZ] reserved for use by xspin interface\n");
 	printf("\t-V print version number and exit\n");
 	alldone(1);
@@ -226,9 +253,34 @@ optimizations(char nr)
 	}
 }
 
+#if 0
+static int
+Rename(const char *old, char *new)
+{	FILE *fo, *fn;
+	char buf[1024];
+
+	if ((fo = fopen(old, "r")) == NULL)
+	{	printf("spin: cannot open %s\n", old);
+		return 1;
+	}
+	if ((fn = fopen(new, "w")) == NULL)
+	{	printf("spin: cannot create %s\n", new);
+		fclose(fo);
+		return 2;
+	}
+	while (fgets(buf, 1024, fo))
+		fputs(buf, fn);
+
+	fclose(fo);
+	fclose(fn);
+
+	return 0;	/* success */
+}
+#endif
+
 int
 main(int argc, char *argv[])
-{	Symbol *s; int preprocessonly = 0;
+{	Symbol *s;
 	int T = (int) time((time_t *)0);
 	int usedopts = 0;
 	extern void ana_src(int, int);
@@ -244,14 +296,15 @@ main(int argc, char *argv[])
 		/* generate code for separate compilation: S1 or S2 */
 		case 'S': separate = atoi(&argv[1][2]);
 			  /* fall through */
-
 		case 'a': analyze  = 1; break;
+
+		case 'A': export_ast = 1; break;
 		case 'B': no_wrapup = 1; break;
 		case 'b': no_print = 1; break;
 		case 'C': Caccess = 1; break;
 		case 'c': columns = 1; break;
 		case 'D': PreArg[++PreCnt] = (char *) &argv[1][0];
-			  break;
+			  break;	/* define */
 		case 'd': dumptab =  1; break;
 		case 'E': PreArg[++PreCnt] = (char *) &argv[1][2];
 			  break;
@@ -260,7 +313,9 @@ main(int argc, char *argv[])
 		case 'f': add_ltl = (char **) argv;
 			  argc--; argv++; break;
 		case 'g': verbose +=  1; break;
+		case 'h': seedy = 1; break;
 		case 'i': interactive = 1; break;
+		case 'I': inlineonly = 1; break;
 		case 'J': like_java = 1; break;
 		case 'j': jumpsteps = atoi(&argv[1][2]); break;
 		case 'l': verbose +=  2; break;
@@ -278,16 +333,20 @@ main(int argc, char *argv[])
 			  break;
 		case 'r': verbose +=  8; break;
 		case 's': verbose += 16; break;
+		case 'T': notabs = 1; break;
 		case 't': s_trail  =  1;
 			  if (isdigit(argv[1][2]))
 				ntrail = atoi(&argv[1][2]);
 			  break;
+		case 'U': PreArg[++PreCnt] = (char *) &argv[1][0];
+			  break;	/* undefine */
+		case 'u': cutoff = atoi(&argv[1][2]); break;	/* new 3.4.14 */
 		case 'v': verbose += 32; break;
 		case 'V': printf("%s\n", Version);
 			  alldone(0);
 			  break;
 		case 'w': verbose += 64; break;
-		case 'X': xspin = 1;
+		case 'X': xspin = notabs = 1;
 #ifndef PC
 			  signal(SIGPIPE, alldone); /* not posix... */
 #endif
@@ -316,20 +375,13 @@ main(int argc, char *argv[])
 	}
 	if (argc > 1)
 	{	char cmd[128], out2[64];
-#ifdef PC
-		strcpy(out1, "_tmp1_");
-		strcpy(out2, "_tmp2_");
-#else
-		/* extern char *tmpnam(char *);  in stdio.h */
+
+		/* must remain in current dir */
+		strcpy(out1, "pan.pre");
+
 		if (add_ltl || nvr_file)
-		{	/* must remain in current dir */
-			strcpy(out1, "_tmp1_");
-			strcpy(out2, "_tmp2_");
-		} else
-		{	(void) tmpnam(out1);
-			(void) tmpnam(out2);
-		}
-#endif
+			strcpy(out2, "pan.___");
+
 		if (add_ltl)
 		{	tl_out = cpyfile(argv[1], out2);
 			nr_errs = tl_main(2, add_ltl);	/* in tl_main.c */
@@ -353,13 +405,8 @@ main(int argc, char *argv[])
 			preprocess(argv[1], out1, 0);
 
 		if (preprocessonly)
-		{	(void) unlink("pan.pre"); /* remove old version */
-			if (rename((const char *) out1, "pan.pre") != 0)
-			{	printf("spin: rename %s failed\n", out1);
-				alldone(1);
-			}
 			alldone(0);
-		}
+
 		if (!(yyin = fopen(out1, "r")))
 		{	printf("spin: cannot open %s\n", out1);
 			alldone(1);
@@ -372,7 +419,7 @@ main(int argc, char *argv[])
 			strcpy(cmd, argv[1]);
 		oFname = Fname = lookup(cmd);
 		if (oFname->name[0] == '\"')
-		{	int i = strlen(oFname->name);
+		{	int i = (int) strlen(oFname->name);
 			oFname->name[i-1] = '\0';
 			oFname = lookup(&oFname->name[1]);
 		}
@@ -401,19 +448,30 @@ main(int argc, char *argv[])
 	if (columns == 2 && limited_vis)
 		verbose += (1+4);
 	Srand(T);	/* defined in run.c */
-	s = lookup("_");	s->type = PREDEF; /* a write only global var */
+	SeedUsed = T;
+	s = lookup("_");	s->type = PREDEF; /* write-only global var */
 	s = lookup("_p");	s->type = PREDEF;
 	s = lookup("_pid");	s->type = PREDEF;
 	s = lookup("_last");	s->type = PREDEF;
 	s = lookup("_nr_pr");	s->type = PREDEF; /* new 3.3.10 */
+
 	yyparse();
 	fclose(yyin);
+	loose_ends();
+
+	if (inlineonly)
+	{	repro_src();
+		return 0;
+	}
+
 	chanaccess();
-	if (!s_trail && (dataflow || merger))
-		ana_src(dataflow, merger);
-	sched();
-	alldone(nr_errs);
-	return 0;	/* can't get here */
+	if (!Caccess)
+	{	if (!s_trail && (dataflow || merger))
+			ana_src(dataflow, merger);
+		sched();
+		alldone(nr_errs);
+	}
+	return 0;
 }
 
 int
@@ -463,7 +521,7 @@ emalloc(int n)
 }
 
 void
-trapwonly(Lextok *n, char *s)
+trapwonly(Lextok *n, char *unused)
 {	extern int realread;
 	short i = (n->sym)?n->sym->type:0;
 
@@ -502,6 +560,7 @@ setaccess(Symbol *sp, Symbol *what, int cnt, int t)
 Lextok *
 nn(Lextok *s, int t, Lextok *ll, Lextok *rl)
 {	Lextok *n = (Lextok *) emalloc(sizeof(Lextok));
+	static int warn_nn = 0;
 
 	n->ntyp = (short) t;
 	if (s && s->fn)
@@ -558,6 +617,9 @@ nn(Lextok *s, int t, Lextok *ll, Lextok *rl)
 			}	}
 			forbidden = 0;
 			break;
+		case 'c':
+			AST_track(n, 0);	/* register as a slice criterion */
+			/* fall thru */
 		default:
 			forbidden = 0;
 			break;
@@ -566,27 +628,62 @@ nn(Lextok *s, int t, Lextok *ll, Lextok *rl)
 		{	printf("spin: never, saw "); explain(t); printf("\n");
 			fatal("incompatible with separate compilation",(char *)0);
 		}
-	} else if (t == ENABLED || t == PC_VAL)
-	{	printf("spin: Warning, using %s outside never-claim\n",
+	} else if ((t == ENABLED || t == PC_VAL) && !(warn_nn&t))
+	{	printf("spin: Warning, using %s outside never claim\n",
 			(t == ENABLED)?"enabled()":"pc_value()");
+		warn_nn |= t;
 	} else if (t == NONPROGRESS)
-	{	fatal("spin: Error, using np_ outside never-claim\n", (char *)0);
+	{	fatal("spin: Error, using np_ outside never claim\n", (char *)0);
 	}
 	return n;
 }
 
 Lextok *
-rem_lab(Symbol *a, Lextok *b, Symbol *c)
+rem_lab(Symbol *a, Lextok *b, Symbol *c)	/* proctype name, pid, label name */
 {	Lextok *tmp1, *tmp2, *tmp3;
 
 	has_remote++;
-	fix_dest(c, a);	/* in case target is jump */
+	c->type = LABEL;	/* refered to in global context here */
+	fix_dest(c, a);		/* in case target of rem_lab is jump */
 	tmp1 = nn(ZN, '?',   b, ZN); tmp1->sym = a;
 	tmp1 = nn(ZN, 'p', tmp1, ZN);
 	tmp1->sym = lookup("_p");
 	tmp2 = nn(ZN, NAME,  ZN, ZN); tmp2->sym = a;
 	tmp3 = nn(ZN, 'q', tmp2, ZN); tmp3->sym = c;
 	return nn(ZN, EQ, tmp1, tmp3);
+#if 0
+	      .---------------EQ-------.
+	     /                          \
+	   'p' -sym-> _p               'q' -sym-> c (label name)
+	   /                           /
+	 '?' -sym-> a (proctype)     NAME -sym-> a (proctype name)
+	 / 
+	b (pid expr)
+#endif
+}
+
+Lextok *
+rem_var(Symbol *a, Lextok *b, Symbol *c, Lextok *ndx)
+{	Lextok *tmp1;
+
+	has_remote++;
+	has_remvar++;
+	dataflow = 0;	/* turn off dead variable resets 4.2.5 */
+	tmp1 = nn(ZN, '?', b, ZN); tmp1->sym = a;
+	tmp1 = nn(ZN, 'p', tmp1, ndx);
+	tmp1->sym = c;
+	return tmp1;
+#if 0
+	cannot refer to struct elements
+	only to scalars and arrays
+
+	    'p' -sym-> c (variable name)
+	    / \______  possible arrayindex on c
+	   /
+	 '?' -sym-> a (proctype)
+	 / 
+	b (pid expr)
+#endif
 }
 
 static void
@@ -594,8 +691,8 @@ explain(int n)
 {	FILE *fd = stdout;
 	switch (n) {
 	default:	if (n > 0 && n < 256)
-				fprintf(fd, "%c' = '", n);
-			fprintf(fd, "%d", n);
+				fprintf(fd, "'%c' = '", n);
+			fprintf(fd, "%d'", n);
 			break;
 	case '\b':	fprintf(fd, "\\b"); break;
 	case '\t':	fprintf(fd, "\\t"); break;
@@ -614,10 +711,16 @@ explain(int n)
 	case ASSERT:	fprintf(fd, "%sassert",	Function); break;
 	case ATOMIC:	fprintf(fd, "%satomic",	Keyword); break;
 	case BREAK:	fprintf(fd, "%sbreak",	Keyword); break;
+	case C_CODE:	fprintf(fd, "%sc_code",	Keyword); break;
+	case C_DECL:	fprintf(fd, "%sc_decl",	Keyword); break;
+	case C_EXPR:	fprintf(fd, "%sc_expr",	Keyword); break;
+	case C_STATE:	fprintf(fd, "%sc_state",Keyword); break;
+	case C_TRACK:	fprintf(fd, "%sc_track",Keyword); break;
 	case CLAIM:	fprintf(fd, "%snever",	Keyword); break;
 	case CONST:	fprintf(fd, "a constant"); break;
 	case DECR:	fprintf(fd, "%s--",	Operator); break;
 	case D_STEP:	fprintf(fd, "%sd_step",	Keyword); break;
+	case D_PROCTYPE: fprintf(fd, "%sd_proctype", Keyword); break;
 	case DO:	fprintf(fd, "%sdo",	Keyword); break;
 	case DOT:	fprintf(fd, "."); break;
 	case ELSE:	fprintf(fd, "%selse",	Keyword); break;
@@ -630,11 +733,13 @@ explain(int n)
 	case GE:	fprintf(fd, "%s>=",	Operator); break;
 	case GOTO:	fprintf(fd, "%sgoto",	Keyword); break;
 	case GT:	fprintf(fd, "%s>",	Operator); break;
+	case HIDDEN:	fprintf(fd, "%shidden",	Keyword); break;
 	case IF:	fprintf(fd, "%sif",	Keyword); break;
 	case INCR:	fprintf(fd, "%s++",	Operator); break;
 	case INAME:	fprintf(fd, "inline name"); break;
 	case INLINE:	fprintf(fd, "%sinline",	Keyword); break;
 	case INIT:	fprintf(fd, "%sinit",	Keyword); break;
+	case ISLOCAL:	fprintf(fd, "%slocal",  Keyword); break;
 	case LABEL:	fprintf(fd, "a label-name"); break;
 	case LE:	fprintf(fd, "%s<=",	Operator); break;
 	case LEN:	fprintf(fd, "%slen",	Function); break;
@@ -655,15 +760,20 @@ explain(int n)
 	case PC_VAL:	fprintf(fd, "%spc_value",Function); break;
 	case PNAME:	fprintf(fd, "process name"); break;
 	case PRINT:	fprintf(fd, "%sprintf",	Function); break;
+	case PRINTM:	fprintf(fd, "%sprintm",	Function); break;
+	case PRIORITY:	fprintf(fd, "%spriority", Keyword); break;
 	case PROCTYPE:	fprintf(fd, "%sproctype",Keyword); break;
+	case PROVIDED:	fprintf(fd, "%sprovided",Keyword); break;
 	case RCV:	fprintf(fd, "%s?",	Operator); break;
 	case R_RCV:	fprintf(fd, "%s??",	Operator); break;
 	case RSHIFT:	fprintf(fd, "%s>>",	Operator); break;
 	case RUN:	fprintf(fd, "%srun",	Operator); break;
 	case SEP:	fprintf(fd, "token: ::"); break;
 	case SEMI:	fprintf(fd, ";"); break;
+	case SHOW:	fprintf(fd, "%sshow", Keyword); break;
 	case SND:	fprintf(fd, "%s!",	Operator); break;
 	case STRING:	fprintf(fd, "a string"); break;
+	case TRACE:	fprintf(fd, "%strace", Keyword); break;
 	case TIMEOUT:	fprintf(fd, "%stimeout",Keyword); break;
 	case TYPE:	fprintf(fd, "data typename"); break;
 	case TYPEDEF:	fprintf(fd, "%stypedef",Keyword); break;
